Code for augmentation and testing augmentation
Added code to augmentation.py in src directory and created a test_augmentation.py in the test directory.

Transformations currently included are color jitter, random rotation and horizontal flip.

Created a new test\res_aug to save the transformations

included a label directory in the test\res, the test\res_out and the test\res_aug to save replica transformations on the masks

Still require to work on:
- add additional transformation and decide on probability factor and any limits
- use to augment the dataset ( at the moment just saves one transformed image)
- look at masks output from transformation to ensure we preserve the 0, 1, 2 and 255 labels