Create: a first pass of ClipSeg model
A rough first pass with all the components of the ClipSeg model.  Likely to be overlap with existing code (e.g. data loading and preprocessing) so expect it can be cut down.  Suggest separating out encoder and decoder in to separate files but for ease of first pass and assessment that we have all required elements I've kept in the one file.