Create: Adds a blurred dot at centre of mask on image
Adds a blurred dot at centre of mask on image (red for cats and green for dogs) and saves in a new directory (/train_centre/).  Also saves masks with blurred dots added in new directory (/label_centre).  The latter is only for visual checking .  Dot size was set t o5 pixels with Gaussian blur set to 9 pixel radius (just assumed values after visual inspections to try).   At present the image is reduced in intensity by 5% so the dots are clear even when the image pixels are white at he centre.  All images are reduced in intensity which may impact training so we may want to try with and without.  A small percentage of  images either do not have masks so no centroid can be calculated or the calculated centroid is outside of the mask.   For these cases I just return the original image with no dot.